Moved spawning functions into seperate file
